Move policy-collection test cases to integration
