+++ test/integration/policy_iam_test.go
@@ -1,4 +1,4 @@
-package e2e
+package integration
 
 import (
 	"context"
